+++ R/apply_class.R
@@ -1,7 +1,7 @@
 #' Apply variable class
 #'
-#' Ensure that fiscal year and targets are numeric and all other variables are
-#' stored as characters.
+#' Ensure that fiscal year, cumulative, and targets are numeric and all other
+#' variables are stored as characters.
 #'
 #' @param df dataframe output to reorder
 #' @export
@@ -10,6 +10,6 @@ apply_class <- function(df){
   #ensure variables are same class (problem when missing all targets)
   df <- df %>%
     dplyr::mutate(dplyr::across(dplyr::everything(), as.character),
-                  dplyr::across(c(fiscal_year, targets), as.double))
+                  dplyr::across(c(fiscal_year, cumulative, targets), as.double))
   return(df)
 }