+++ R/limit_datatype.R
@@ -12,12 +12,14 @@
 
 limit_datatype <- function(df, type){
 
-  if("data_type" %in% names(df)){
-    #identify what to keep
-    keep <- ifelse(type == "PLHIV", "SUBNAT", "MER")
+  if("data_type" %in% names(df) && type == "PLHIV"){
+    #limit data to SUBNAT or IMPATT
+    df <- dplyr::filter(df, data_type %in% c("SUBNAT", "IMPATT"))
+  }
 
-    #limit data to SUBNAT or MER
-    df <- dplyr::filter(df, data_type == keep)
+  if("data_type" %in% names(df) && type != "PLHIV"){
+    #limit data to MER
+    df <- dplyr::filter(df, data_type == "MER")
   }
 
   return(df)