+++ R/utilities.R
@@ -108,8 +108,8 @@ return_tab <- function(type){
 
 #' Apply Fiscal Year
 #'
-#' Apply fiscal year to each row, using the T or T_1 in `indicator_code` to
-#' determine whether it's the current or prior fiscal year. The fiscal year
+#' Apply fiscal year to each row, using the T or T_1 or R in `indicator_code` to
+#' determine whether it's the current or  a prior fiscal year. The fiscal year
 #' can be identified dynamically through `grab_info()`.
 #'
 #' @param df DP dataframe to apply fiscal year to
@@ -122,10 +122,46 @@ apply_fy <- function(df, year){
 
   df <- df %>%
     dplyr::mutate(fiscal_year =
-                    ifelse(stringr::str_detect(indicator_code, "\\.(T_1)$"),
-                           year-1,
-                           year)
+                    dplyr::case_when(stringr::str_detect(indicator_code, "\\.(T_1)$") ~ year-1,
+                                     stringr::str_detect(indicator_code, "\\.R$") ~ year-2,
+                                     TRUE ~ year)
     )
 
   return(df)
 }
+
+
+#' Identify Prioritization
+#'
+#' Pull from the prioritization tab to have a table of PSNU prioritization for the
+#' current COP.
+#'
+#' @param filepath file path to the Data Pack importing, must be .xlsx
+#'
+#' @return dataframe from the Prioritization tab
+#' @export
+#' @family prioritization
+
+grab_prioritization <- function(filepath){
+  import_dp(filepath, "Prioritization") %>%
+    split_psnu() %>%
+    dplyr::select(psnuuid, snuprioritization)
+}
+
+#' Apply Prioritization
+#'
+#' Join the new COP prioritization onto the target data frame.
+#'
+#' @param filepath file path to the Data Pack importing, must be .xlsx
+#'
+#' @return df data pack data frame
+#' @return df_prioritization dataframe from `grab_prioritization()`
+#' @export
+#' @family prioritization
+
+apply_prioritization <- function(df, df_prioritization){
+  df %>%
+    dplyr::left_join(df_prioritization, by = c("psnuuid")) %>%
+    dplyr::relocate(snuprioritization, .after = "psnuuid")
+}
+